+++ ESPixelStick/src/output/OutputCommon.cpp
@@ -122,7 +122,7 @@ void c_OutputCommon::GetConfig (JsonObject & jsonConfig)
 } // GetConfig
 
 //----------------------------------------------------------------------------
-void c_OutputCommon::WriteChannelData (size_t StartChannelId, size_t ChannelCount, byte * pSourceData)
+void c_OutputCommon::WriteChannelData (uint32_t StartChannelId, uint32_t ChannelCount, byte * pSourceData)
 {
     // DEBUG_START;
 
@@ -135,7 +135,7 @@ void c_OutputCommon::WriteChannelData (size_t StartChannelId, size_t ChannelCoun
 } // WriteChannelData
 
 //----------------------------------------------------------------------------
-void c_OutputCommon::ReadChannelData(size_t StartChannelId, size_t ChannelCount, byte * pTargetData)
+void c_OutputCommon::ReadChannelData(uint32_t StartChannelId, uint32_t ChannelCount, byte * pTargetData)
 {
     // DEBUG_START;
